Add files via upload
Finally found the secret trigger that makes it all work. Using this trigger, onUserActive, eliminates the need to react to three other individual triggers, simplifying the code and making it work as it should. Auto-subscription now sets the profile field checkboxes it is supposed to check right after the user registration process is completed.